avoid Visual C++ warnings
git-svn-id: svn://hevc.kw.bbc.co.uk/svn/jctvc-tmuc/branches/HM-3.2-dev@1058 ce97f6c2-8f86-0410-b44d-af84b13ec34f

+++ source/Lib/TLibCommon/TComDataCU.cpp
@@ -172,8 +172,8 @@ Void TComDataCU::create(UInt uiNumPartition, UInt uiWidth, UInt uiHeight, Bool b
   }
   
 #if FINE_GRANULARITY_SLICES
-  m_uiSliceStartCU        = (UInt*  )xMalloc(UInt, (1<<(g_uiMaxCUDepth<<1)));
-  m_uiEntropySliceStartCU = (UInt*  )xMalloc(UInt, (1<<(g_uiMaxCUDepth<<1)));
+  m_uiSliceStartCU        = (UInt*  )xMalloc(UInt, ((size_t)1<<(g_uiMaxCUDepth<<1)));
+  m_uiEntropySliceStartCU = (UInt*  )xMalloc(UInt, ((size_t)1<<(g_uiMaxCUDepth<<1)));
 #endif
   // create pattern memory
   m_pcPattern            = (TComPattern*)xMalloc(TComPattern, 1);
